backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
httpx==0.27.0
passlib[bcrypt]==1.7.4


alembic==1.13.2
Jinja2==3.1.4
python-multipart==0.0.9
